push add project page fixed bug on delete task id check ok clean up some lines made folders for handle submit project task and task roadmap put alertgialog jsx in services folder
