+++ frontend/src/pages/AddProject/AddProject.jsx
@@ -1,18 +1,14 @@
 import React, { useState } from "react";
-import { Box } from "@mui/material";
-import AddTaskForm from "@components/TodoList/HandleSubmitTask";
-import AddProjectForm from "../../components/TodoList/HandleSubmitProject";
+import { Box, useTheme } from "@mui/material";
+import AddTaskForm from "@components/HandleSumitTask/HandleSubmitTask";
+import AddProjectForm from "../../components/HandleSubmitProject/HandleSubmitProject";
 // import AddTaskCard from "../../components/TodoList/HandleSubmitTest";
-
-// import { tokens } from "../../theme";
-// import Input from "@mui/material/Input";
-// import ToDoList from "../../components/TodoList/TodoList";
+import { tokens } from "../../theme";
 import Header from "../../components/Header/Header";
 
 function AddProject() {
-  // const theme = useTheme();
-  // const colors = tokens(theme.palette.mode);
-  // const [startDate, setStartDate] = useState("");
+  const theme = useTheme();
+  const colors = tokens(theme.palette.mode);
   const [projectName, setProjectName] = useState("");
 
   return (
@@ -25,7 +21,7 @@ function AddProject() {
           justifyContent="center"
           borderRadius={1}
           p={1}
-          sx={{ backgroundColor: "#f2f0f0" }}
+          sx={{ backgroundColor: colors.grey[500] }}
         >
           <AddProjectForm setProjectName={setProjectName} />
         </Box>
@@ -41,18 +37,6 @@ function AddProject() {
         >
           <AddTaskForm projectName={projectName} />
         </Box>
-
-        {/* <Box
-          display="flex"
-          m={1}
-          flexDirection="column"
-          justifyContent="center"
-          borderRadius={1}
-          p={1}
-          sx={{ backgroundColor: "#f2f0f0" }}
-        >
-          <AddTaskForm projectName={projectName} />
-        </Box>  */}
       </Box>
     </Box>
   );
