+++ frontend/src/pages/Roadmap/Roadmap.jsx
@@ -1,9 +1,9 @@
 import React from "react";
-// import Task from "@components/Task/Task";
+import Task from "@components/Task/Task";
 import { Box } from "@mui/material";
 import Header from "@components/Header/Header";
 // import AddTaskForm from "@components/TodoList/HandleSubmitTask";
-import AddTaskRoadmap from "@components/TodoList/HandleSubmitTaskRoadmap";
+import AddTaskRoadmap from "@components/HandleSubmitTaskRoadmap/HandleSubmitTaskRoadmap";
 
 function Roadmap() {
   return (
@@ -11,6 +11,7 @@ function Roadmap() {
       <Header title="ROADMAP" subtitle="Tasks list" />
       <Box marginBottom={1}>
         <AddTaskRoadmap />
+        <Task />
       </Box>
     </Box>
   );
